SDK, analytics: Some warnings fixed
<==== Mercurial data ====>
Branch: meta_4.0

+++ vms/libs/nx_sdk/src/nx/sdk/analytics/helpers/event_metadata_packet.cpp
@@ -39,7 +39,7 @@ int EventMetadataPacket::count() const
 
 const IEventMetadata* EventMetadataPacket::at(int index) const
 {
-    if (index < 0 || index >= m_events.size())
+    if (index < 0 || index >= (int) m_events.size())
         return nullptr;
 
     return m_events[index];
